Encapsulate connection establishing code in a decorator
Code that establishes connection with the database and
creates a table was split in two parts. The connection
establishments now occurs in a decorator (connect). The "create_table" funcion, as well as other functions requiring connection with the database that will be developed in the future, will be decorated by that decorator